+++ gakumas-tools/components/IdolIcon.js
@@ -1,0 +1,17 @@
+import { Idols } from "gakumas-data";
+import gkImg from "gakumas-images";
+import Image from "@/components/Image";
+
+export default function IdolIcon({ idolId }) {
+  const idol = Idols.getById(idolId);
+  return (
+    <Image
+      style={{ verticalAlign: "sub" }}
+      src={gkImg(idol).icon}
+      alt=""
+      width={24}
+      height={24}
+      draggable={false}
+    />
+  );
+}