+++ gakumas-tools/components/Oshi/Oshi.js
@@ -1,21 +1,20 @@
 "use client";
 import { useState } from "react";
 import { Resizable } from "re-resizable";
+import IdolIcon from "@/components/IdolIcon";
 import YouTubeVideo from "@/components/YouTubeVideo";
 import styles from "./Oshi.module.scss";
 
-const Badge = () => (
-  <div
-    style={{
-      display: "inline-block",
-      marginRight: 8,
-      backgroundColor: "#ff3333",
-      borderRadius: 5,
-      width: 10,
-      height: 10,
-    }}
-  />
-);
+export const OSHI_PROPS = {
+  text: (
+    <div>
+      <IdolIcon idolId={6} />
+      「伊藤舞音のおいでよ まおーこくだい9かい」
+    </div>
+  ),
+  initiallyExpanded: true,
+  url: "https://www.youtube.com/watch?v=XDT1HNaCAeA",
+};
 
 export default function Oshi({
   text,
@@ -41,7 +40,7 @@ export default function Oshi({
           </a>
         ) : (
           <button className={styles.expand} onClick={() => setExpanded(true)}>
-            {hasBadge && <Badge />}{" "}
+            {hasBadge && <div className={styles.badge} />}{" "}
           </button>
         )}
       </div>