+++ gakumas-tools/components/Simulator/Simulator.js
@@ -204,6 +204,7 @@ export default function Simulator() {
     <div id="simulator_loadout" className={styles.loadoutEditor}>
       <div className={styles.configurator}>
         <div>{t("multiplierNote")}</div>
+        <div>※ Some card effects may be inaccurate.</div>
         {stage.preview && <div>{t("previewNote")}</div>}
         <StageSelect />
         {stage.type !== "contest" ? (
@@ -299,7 +300,7 @@ export default function Simulator() {
             href="https://github.com/surisuririsu/gakumas-tools/blob/master/gakumas-tools/simulator/CHANGELOG.md"
             target="_blank"
           >
-            {t("lastUpdated")}: 2025-12-20
+            {t("lastUpdated")}: 2025-12-26
           </a>
         </div>
         {!simulatorData && (